+++ src/main/kotlin/io/sdkman/dto/ErrorResponse.kt
@@ -1,0 +1,9 @@
+package io.sdkman.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ErrorResponse(
+    val error: String,
+    val message: String,
+)
