+++ src/main/kotlin/io/sdkman/dto/HealthCheckResponse.kt
@@ -1,0 +1,15 @@
+@file:UseSerializers(OptionSerializer::class)
+
+package io.sdkman.dto
+
+import arrow.core.None
+import arrow.core.Option
+import arrow.core.serialization.OptionSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.UseSerializers
+
+@Serializable
+data class HealthCheckResponse(
+    val status: String,
+    val message: Option<String> = None,
+)
